Support Vault autodiscovery and JWT auth
This also allows setting a pre-environment file for adding variables to
the process environment before it runs as well as specifying the target
directory as an environment variable. The goal of these changes is to
allow this binary to by symlinked directly into the cron daily directory
rather than needing a bin-stub script to start it.